refactor(examples/ai-collab): Simplify and fix a few things (#22807)
## Description

- The URL for the ai-collab sample app is now `http://localhost:3000`
instead of `http://localhost:3000/tasks-list`, for simplicity. The
README had a typo so it pointed to a bad URL, so that has been fixed to
point to the new simplified URL too.
- Adds a check before using `window` to account for NextJS's server-side
rendering, to avoid an error in the console output when running, and
letting `next build` complete successfully.
- Removes NextJS-related npm scripts that would be important in a prod
application but we don't really use in the sample app.