Removed the singleton implementation of the startup probe (#22819)
## Description

The singleton implementation of `StartupCheck` causes bugs when the r11
packages do not match the historian packages consumed. Hence, I decided
to switch to a non-singleton implementation. This introduces the
`StartupCheck` as an implementation of `IReadinessCheck`. This probe is
a resource provided to all HTTP services in r11.

## Breaking Changes

Changes Resource and Runner objects for Alfred, Riddler and Nexus to
include the `StartUpCheck` object.

---------

Co-authored-by: Tyler Butler <tyler@tylerbutler.com>