Removed singleton Startup check usage by Historian and Gitrest (#22826)
## Description

This PR follows the r11s PR -
https://github.com/microsoft/FluidFramework/pull/22819 - to remove the
usage of the `StartupCheck` singleton by Historian and Gitrest. This
singleton along with r11 package mismatch caused bugs. Hence, now I pass
the implementation of the startup probe as a resource to the server.

## Breaking Changes

Changes the resourceFactory args of both Gitrest and Historian.

---------

Co-authored-by: Tyler Butler <tyler@tylerbutler.com>