When trimming the EditManager trunk, avoid mutating revision and change of the new trunk base commit (#22869)
## Description

This gets rid of unnecessary mutations of the commit that becomes the
trunk base commit during trunk trimming in the EditManager. The mutation
of the `revision` property was done to avoid an additional check in the
`getSummaryData()` codepath, and the mutation of the `change` property
was done merely for consistency. This gets rid of both in favor of just
doing the simple check in `getSummaryData()`, since it is best to avoid
mutating objects that are otherwise considered to be immutable.

This PR also:
* adds a test for a scenario that is currently failing - namely,
reverting a commit (e.g. via an undo) when the SharedTree is detached.
This will be fixed in a follow-up PR.
* adds documentation for the special `"root"` revision tag.