Enabled support for a token issuance endpoint in Alfred (#22884)
## Description

This PR adds support for a fluid token issuance endpoint. This endpoint
can be used to issue fluid access tokens based on a custom
implementation of the `IFluidAccessTokenGenerator` interface. One use of
this endpoint is to enable access using cloud identity providers such as
Entra-ID. Alfred has a new endpoint -
`api/v1/tenants/:tenantid/accesstoken`. This endpoint expects a `Bearer`
token and creates an access token for a given fluid tenant.

This PR also adds support to inject a custom implementation of the
`IFluidAccessTokenGenerator`. This can be used to implement any custom
logic that is needed for a business use case.

It also has unit tests for the following:

1) Throttling of the endpoint
2) Validation of cases where a `Bearer` token is not provided or an
invalid authorization method is used
3) Validation of a valid token creation path
4) Validation of token creation failure due to invalid token signature
and due to unauthorized access

## Breaking Changes

This PR adds a new resourceFactory arg of type
`IFluidAccessTokenGenerator`. This arg is customizable as well.

---------

Co-authored-by: Tyler Butler <tyler@tylerbutler.com>