build(build-tools): Upgrade jssm and jssm-viz, patch jssm-viz (#22880)
The latest release of jssm includes the fixes in
https://github.com/StoneCypher/jssm/pull/569, so our patch is no longer
needed and has been removed. However, the jssm-viz package has a similar
problem (open PR: https://github.com/StoneCypher/jssm-viz/pull/54) so I
patched it with type changes similar to what was done for jssm
originally.

If or when the jssm-viz PR is accepted or the issue is otherwise fixed,
we can remove the patch.