Fix handle resolution bug and add test for synchronous datastore creation (#22917)
Adding a test that attempts to create a new datastore as synchronously
as possible. In the process of creating this test, i noticed a bug in
the handle resolution code, where handles to datastores would fail to
resolve, unless customer request logic, like in data object, is used. It
should not be necessary to use custom request logic, as it is pattern
the runtime has been moving away from. To fix this I've added handling
to the datastore runtime such that it returns the entrypoint when the
request is via a handle, and has no sub-path.