Include service URL and request method in r11s request errors (#22851)
Troubleshooting live-site issues and customer reported issues will
become easier. Ex: If Container creation failures have service URL and
response code, it'll be easier to identify point of failure and come-up
with a fix to mitigate/fix the incident.

This gives AFR a good idea about the APIs which are failing, which
cluster it was going to, and which cluster group it was going to. It
helps AFR correlate the errors on the server better.

## Examples

Here are some examples of request calls. Only the bolded parts can be
logged, and the first line of each entry below is just the type of
request being made (no need to log it):

**CreateDocument** (or **CreateContainer**)
**POST** https://**us.fluidrelay.azure.com/documents**/someId

**Connect**  
**GET**
https://**alfred.southcentralus-1.fluidrelay.azure.com/socket.io**/?documentId=someId&tenantId=someId&EIO=4&transport=websocket
 
**GetDelta**  
**GET**
https://**alfred.southcentralus-1.fluidrelay.azure.com/deltas**/someId/someId?from=0&to=2001

**GetSummary**  
**GET**
https://**historian.southcentralus-1.fluidrelay.azure.com/repos**/someId/git/summaries/latest?token=someToken&disableCache=true

**PostSummary**  
**POST**
https://**historian.southcentralus-1.fluidrelay.azure.com/repos**/someId/git/summaries?token=someToken

**FindSession** – **First Request**
**POST**
https://**alfred.southcentralus-1.fluidrelay.azure.com/documents**/someId/session/someId

Fixes
[AB#18452](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/18452)