fix(release): Check release notes and changelog generation in release tool (#22811)
Adds steps to the release tool to check that release notes have been
generated and that the per-package changelogs have been updated. Also
prompts the user for a bump type earlier in the process if needed.

The changelog step is not trivial to check, so I just prompt the user to
select whether they did it. Not ideal, but at least the tool will remind
the release driver that this step is needed. We can add a more complete
check in the future if needed.

---------

Co-authored-by: Joshua Smithrud <54606601+Josmithr@users.noreply.github.com>