Delay edit manager trunk eviction when detached (#22930)
## Description

When a SharedTree is detached and submits a change, this schedules the
subsequent advancement of the EditManager's minimum sequence number on
the JS microtask queue rather than doing it immediately/synchronously.
Doing it synchronously is problematic because advancing the minimum
sequence number might cause trunk commits to be evicted, however, this
is happening in an event callback context - `submitCommit` is being
called as a result of the `afterChange` event being fired on the local
branch. If trunk eviction were to occur immediately, it would mean that
other listeners to the `afterChange` event might experience different
trunks (some before trimming, others after trimming). In practice, this
is a problem when generating revertibles, as revertibles respond to the
`afterChange` event by creating branches - but the bases of these
branches may or may not have already been trimmed by the aforementioned
behavior. Delaying the trimming guarantees that all listeners to the
`afterChange` event can safely reference any commits that they may
already have handles to.