refactor(build-infrastructure): Rename FluidRepo to BuildProject (#22967)
There is nothing particularly Fluid-centric about the top-level API in
build-infrastructure; I just couldn't think of a better name at the
time. After discussion, we're going to call the top-level object a
BuildProject. The full hierarchy is now:

BuildProjects contain Workspaces contain ReleaseGroups contain Packages

This change required a lot of documentation updates and renaming other
related properties and functions beyond the BuildProject itself.

The package is private so these changes are safe.