refactor(eslint-config-fluid): Disable formatting-related rules (#22959)
Over the years eslint and plugins have gradually deprecated many of
their formatting-related rules in favor of dedicated formatting tools
like biome and prettier. Since we already have a dedicated formatting
tool, this change disables the formatting related rules we use. All of
them have been removed in recent eslint and typescript-eslint releases,
and are unnecessary since we use biome or prettier for formatting.

Related to
[AB#9083](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/9083).