Fix race condition while establishing connection due to multiplexing in ODSP Driver (#23060)
## Description


[AB#22192](https://dev.azure.com/fluidframework/internal/_workitems/edit/22192)

This fixes a bug reported by Loop where the reconnecting banner did not
go away after user returned after being dormant for some time.

Reason for the bug:
Due to socket reuse(multiplexing), we can get "disconnect" event from
other clients in the socket reference. So, a race condition could
happen, where this client is establishing connection and listening for
"connect_document_success" on the socket among other events, but we get
"disconnect" event on the socket reference from other clients, in which
case, we dispose connection object and stop listening to further events
on the socket. Due to this we get stuck as the connection is not yet
established and so we don't return any connection object to the client
(connection manager). So, we remain stuck.
In order to handle this, listen for the "disconnect" event and reject
the promise with the error so that the caller can know and handle the
error.

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>